+++ packages/router/rollup.config.cjs
@@ -16,6 +16,7 @@ const banner = `/**
 
 const entries = {
   'vureact-router': 'src/index.ts',
+  'type-compat': 'src/vue-router-type-compat.ts',
 };
 
 const outputItem = (type = 'js', format = 'cjs') => ({